# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: OPPO Factory CMMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /training:
    get:
      operationId: getTrainingPlans
      tags: [training]
      summary: Get training plans
      responses:
        "200":
          description: List of training plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrainingPlan"
    post:
      operationId: createTrainingPlan
      tags: [training]
      summary: Create training plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTrainingPlanRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingPlan"
  /training/{id}:
    get:
      operationId: getTrainingPlanById
      tags: [training]
      summary: Get training plan by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Training plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingPlan"
    put:
      operationId: updateTrainingPlan
      tags: [training]
      summary: Update training plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTrainingPlanRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingPlan"
    delete:
      operationId: deleteTrainingPlan
      tags: [training]
      summary: Delete training plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /tasks:
    get:
      operationId: getTasks
      tags: [tasks]
      summary: Get tasks and notifications
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    put:
      operationId: updateTask
      tags: [tasks]
      summary: Update task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    TrainingPlan:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        trainerName:
          type: string
        scheduledDate:
          type: string
        scheduledTime:
          type: string
        durationMinutes:
          type: integer
        location:
          type: string
        status:
          type: string
          enum: [scheduled, completed, cancelled]
        participants:
          type: array
          items:
            type: integer
        participantNames:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required:
        - id
        - title
        - scheduledDate
        - status
    CreateTrainingPlanRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        trainerName:
          type: string
        scheduledDate:
          type: string
        scheduledTime:
          type: string
        durationMinutes:
          type: integer
        location:
          type: string
        participants:
          type: array
          items:
            type: integer
      required:
        - title
        - scheduledDate
    UpdateTrainingPlanRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        trainerName:
          type: string
        scheduledDate:
          type: string
        scheduledTime:
          type: string
        durationMinutes:
          type: integer
        location:
          type: string
        status:
          type: string
          enum: [scheduled, completed, cancelled]
        participants:
          type: array
          items:
            type: integer
...
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        assignedToId:
          type: integer
        assignedToName:
          type: string
        assignedById:
          type: integer
        assignedByName:
          type: string
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
        dueDate:
          type: string
        completedAt:
          type: string
        type:
          type: string
          enum: [maintenance, inspection, repair, general]
        relatedMachine:
          type: string
        lineId:
          type: integer
        lineName:
          type: string
        createdAt:
          type: string
      required:
        - id
        - title
        - assignedToId
        - priority
        - status
        - type
    CreateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        assignedToId:
          type: integer
        priority:
          type: string
          enum: [low, medium, high, urgent]
        dueDate:
          type: string
        type:
          type: string
          enum: [maintenance, inspection, repair, general]
        relatedMachine:
          type: string
        lineId:
          type: integer
      required:
        - title
        - assignedToId
        - priority
        - type
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        assignedToId:
          type: integer
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
        dueDate:
          type: string
        completedAt:
          type: string
...